requirements.txt
alembic==1.9.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
autopep8==2.0.2
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dateutils==0.6.12
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Assets==2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.39.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pyclean==2.7.0
pycodestyle==2.10.0
pycparser==2.21
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
sqlalchemy-json==0.5.0
stripe==5.2.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
webassets==2.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.14.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
Brotli==1.0.9
cryptography==40.0.2
cssselect2==0.7.0
html5lib==1.1
lxml==4.9.2
oscrypto==1.3.0
pisa==3.0.33
pydyf==0.6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.8.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
svglib==1.5.1
tinycss2==1.2.1
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
webencodings==0.5.1
xhtml2pdf==0.2.10
zopfli==0.2.2